Split PDF import commit
