+++ tests/test-quality-guardrails.js
@@ -89,6 +89,7 @@ const highValueCheckJsModules = [
   'js/lens.js',
   'js/light-devices-runtime.js',
   'js/light-tool-camera-modals.js',
+  'js/pdf-import-commit.js',
   'js/pdf-import.js',
   'js/profile.js',
   'js/profile-runtime.js',
@@ -363,6 +364,7 @@ const pdfReportCheckJsModules = [
   'js/export-report-builder.js',
   'js/export-report-html.js',
   'js/pdf-import-ai-utils.js',
+  'js/pdf-import-commit.js',
   'js/pdf-import-file-utils.js',
   'js/pdf-import-marker-mapping.js',
   'js/pdf-import-marker-normalization.js',
