+++ tsconfig.json
@@ -174,6 +174,7 @@
     "js/nostr-discovery.js",
     "js/onboarding-view-runtime.js",
     "js/onboarding-view.js",
+    "js/pdf-import-commit.js",
     "js/pdf-import.js",
     "js/pdf-import-marker-mapping.js",
     "js/pdf-import-marker-normalization.js",
